Add code to study the effects on knn
If the kNN results are locate in the same image with the query/probe
patches, which may result in overfitting. However, adding constrains
on it may increace computation complexity significantly. So we take
some experiments to study if it is necessary to add constrains when
compute knn.

by da.li on 2017/04/21